Refactor: add tools dir
