+++ packages/core/tools/snapshot.ts
@@ -17,7 +17,7 @@
 import { z } from 'zod';
 import zodToJsonSchema from 'zod-to-json-schema';
 
-import { captureAriaSnapshot, runAndWait } from './utils';
+import { captureAriaSnapshot, runAndWait } from '../utils';
 
 import type * as playwright from 'playwright';
 import type { Tool } from './tool';
