+++ packages/core/tools/tool.ts
@@ -16,7 +16,7 @@
 
 import type { ImageContent, TextContent } from '@modelcontextprotocol/sdk/types';
 import type { JsonSchema7Type } from 'zod-to-json-schema';
-import type { Context } from '../../src/context';
+import type { Context } from '../context';
 
 export type ToolSchema = {
   name: string;
