+++ tsconfig.json
@@ -9,6 +9,6 @@
     "outDir": "./lib"
   },
   "include": [
-    "src", "packages/core/common.ts", "packages/core/screenshot.ts", "packages/core/snapshot.ts", "packages/core/utils.ts", "packages/testickl/endtoend.ts", "packages/core/tool.ts", "packages/mcp/server/index.ts",
+    "src", "packages/core/tools/common.ts", "packages/core/tools/screenshot.ts", "packages/core/tools/snapshot.ts", "packages/core/utils.ts", "packages/testickl/endtoend.ts", "packages/core/tools/tool.ts", "packages/mcp/server/index.ts", "packages/core/context.ts",
   ],
 }
